feat: don't migrate db to production-server AB#6131

+++ tools/deploy.sh
@@ -59,20 +59,22 @@ function deploy() {
     }
 
     function migrate_database() {
-        log "Migrating database..."
+        if [ "$PRODUCTION_DATA_SERVER" = no ]; then
+            log "Migrating database..."
 
-        declare -a arr=("IBF-static-input")
+            declare -a arr=("IBF-static-input")
 
-        for SCHEMA in "${arr[@]}"
-        do
-            echo "$SCHEMA"
-            rm -f tools/db-dumps/ibf_$SCHEMA.dump
-            PGPASSWORD=$DB_PASSWORD pg_dump -U $DB_USERNAME -Fc -f tools/db-dumps/ibf_$SCHEMA.dump -h $DB_HOST -n \"$SCHEMA\" geonode_datav3
-            PGPASSWORD=$DB_PASSWORD psql -U $DB_USERNAME -d $DB_DATABASE -h $DB_HOST -c 'drop schema "'$SCHEMA'" cascade; create schema "'$SCHEMA'";'
-            PGPASSWORD=$DB_PASSWORD pg_restore -U $DB_USERNAME -d $DB_DATABASE -h $DB_HOST --schema=$SCHEMA --clean tools/db-dumps/ibf_$SCHEMA.dump
-        done
+            for SCHEMA in "${arr[@]}"
+            do
+                echo "$SCHEMA"
+                rm -f tools/db-dumps/ibf_$SCHEMA.dump
+                PGPASSWORD=$DB_PASSWORD pg_dump -U $DB_USERNAME -Fc -f tools/db-dumps/ibf_$SCHEMA.dump -h $DB_HOST -n \"$SCHEMA\" geonode_datav3
+                PGPASSWORD=$DB_PASSWORD psql -U $DB_USERNAME -d $DB_DATABASE -h $DB_HOST -c 'drop schema "'$SCHEMA'" cascade; create schema "'$SCHEMA'";'
+                PGPASSWORD=$DB_PASSWORD pg_restore -U $DB_USERNAME -d $DB_DATABASE -h $DB_HOST --schema=$SCHEMA --clean tools/db-dumps/ibf_$SCHEMA.dump
+            done
 
-        PGPASSWORD=$DB_PASSWORD psql -U $DB_USERNAME -d $DB_DATABASE -h $DB_HOST -f $SQL_FILE_PATH
+            PGPASSWORD=$DB_PASSWORD psql -U $DB_USERNAME -d $DB_DATABASE -h $DB_HOST -f $SQL_FILE_PATH
+        fi
 
     }
